Switch ApiListProviderBase to use the hacked "Variables.MainForm.TheSession.Editor.SynchronousEditor.HttpGet" rather than Tools.GetHtml
Makes Watchlist making work, and other providers use their proper user limits on wiki etc (Good for users ;).

It is a bit hacked, we know!

Added an QueryApi method to ApiEdit.. (To be used in the list providers) Not in use yet, as need to find a way for passing in arrays that we'll need to alter/append to as extra parameters and such are added



Whats protected abstract ICollection<string> Actions { get; } for in ApiListProviderBase ?? Is this for generalising (for change to session) stuff? Use or remove/comment?

git-svn-id: https://autowikibrowser.svn.sourceforge.net/svnroot/autowikibrowser/AWB@5207 b999ae8e-1018-0410-9b92-f9ef6f55c712