Followup r7084
Not perfect (needs a bit of tweaking), but we can do page existence by checking the PageInfo cached...

Also, I suppose a category could exist if it's got members, but no text, or a blank page (this fixes that issue).. No idea what the API returns for non existant pages and no members, nothing maybe?

Bit of a todo left here.

git-svn-id: https://autowikibrowser.svn.sourceforge.net/svnroot/autowikibrowser/AWB@7178 b999ae8e-1018-0410-9b92-f9ef6f55c712