Fix VS2012 compiler noise.
Although libraries can define as x86/x64, they're not, and at the mercy of the loader application. Due to compiled regex stupidities (TESTME), in AWB this is forced to x86, so make all the DLLs do the same to shut up the noise

There was a mismatch between the processor architecture of the project being built "MSIL" and the processor architecture of the reference "C:\programming\autowikibrowser\AWB\WikiFunctions\bin\Debug\WikiFunctions.dll", "x86". This mismatch may cause runtime failures. Please consider changing the targeted processor architecture of your project through the Configuration Manager so as to align the processor architectures between your project and references, or take a dependency on references with a processor architecture that matches the targeted processor architecture of your project.


git-svn-id: svn://svn.code.sf.net/p/autowikibrowser/code/AWB@9041 5227b50d-9349-4574-b055-4b0a7525e11c